made get sites and get stewardship activities public
