+++ service/src/main/java/com/codeforcommunity/ServiceMain.java
@@ -6,6 +6,7 @@ import com.codeforcommunity.api.ILeaderboardProcessor;
 import com.codeforcommunity.api.IMapProcessor;
 import com.codeforcommunity.api.IProtectedUserProcessor;
 import com.codeforcommunity.api.IReservationProcessor;
+import com.codeforcommunity.api.IProtectedSiteProcessor;
 import com.codeforcommunity.api.ISiteProcessor;
 import com.codeforcommunity.api.ITeamsProcessor;
 import com.codeforcommunity.auth.JWTAuthorizer;
@@ -18,6 +19,7 @@ import com.codeforcommunity.processor.LeaderboardProcessorImpl;
 import com.codeforcommunity.processor.MapProcessorImpl;
 import com.codeforcommunity.processor.ProtectedUserProcessorImpl;
 import com.codeforcommunity.processor.ReservationProcessorImpl;
+import com.codeforcommunity.processor.ProtectedSiteProcessorImpl;
 import com.codeforcommunity.processor.SiteProcessorImpl;
 import com.codeforcommunity.processor.TeamsProcessorImpl;
 import com.codeforcommunity.propertiesLoader.PropertiesLoader;
@@ -99,6 +101,7 @@ public class ServiceMain {
     ILeaderboardProcessor leaderboardProc = new LeaderboardProcessorImpl(this.db);
     IMapProcessor mapProc = new MapProcessorImpl(this.db);
     ITeamsProcessor teamsProc = new TeamsProcessorImpl(this.db);
+    IProtectedSiteProcessor protectedSiteProc = new ProtectedSiteProcessorImpl(this.db);
     ISiteProcessor siteProc = new SiteProcessorImpl(this.db);
 
     // Create the API router and start the HTTP server
@@ -111,6 +114,7 @@ public class ServiceMain {
             leaderboardProc,
             mapProc,
             teamsProc,
+            protectedSiteProc,
             siteProc,
             jwtAuthorizer);
